Add vlucas/phpdotenv to the composer repo
Magento normally has a ".env" file that contains a bunch of environment
specific information, such as where the database is, what the password
is and the like. The problem is, not all the information there is environment
specific. Also, I don't like the format. Configuration in a PHP array?
Blah. It feels ... unclean.

Enter the .env file!

Here's a file I can absolutely guarantee should never ever be checked
into version control. I can guarantee that because it's sole purpose is
not to be checked into version control. I can look at it with comfort,
pass it to kubernetes to be a managed secret, generally feel smug and
superior about the syntax of my environment file.

Or something. I like it, and I don't like how Magento concatenates
information that is strictly secret with information that isn't.

I haven't done anything with it yet. Who knows, I might not.